refactor(pm): move non-registry cache-slot helpers into resolver::common
`file_cache_slot` / `lexically_normalize` (local-tarball path keying) lived in
`resolver::http`, the HTTP tarball resolver — only because they shared the
`cache_slot` sha256 primitive with `http_cache_slot`. After the resolver split
this left `resolver::file` reaching into `resolver::http` for file logic.

Move all four slot helpers (`cache_slot`, `http_cache_slot`, `file_cache_slot`,
`lexically_normalize`) and their tests into `resolver::common`, next to
`commit_cache_dir_atomic` (the slot-*commit* protocol), so all cache-slot
concerns live in one shared-non-registry-helpers module. `http` and `file`
resolvers now both import from `common`. Rename the lib façade `http` →
`cache_slot` (sole consumer is pm) so the public surface is honest. All helpers
stay gated behind `http-tarball`; the wasm (no-feature) build is unaffected.

Pure relocation — no behavior change. Full ruborist suite + outside-root file:
tarball install verified.
