+++ crates/ruborist/src/resolver/common.rs
@@ -23,6 +23,8 @@ use std::path::Path;
 #[cfg(any(feature = "native-git", feature = "http-tarball"))]
 use anyhow::Context;
 use anyhow::{Result, anyhow};
+#[cfg(feature = "http-tarball")]
+use sha2::{Digest, Sha256};
 
 use crate::model::manifest::{CoreVersionManifest, Dist};
 use crate::util::OnceMap;
@@ -180,9 +182,160 @@ where
         .await
 }
 
+// ---------------------------------------------------------------------------
+// Non-registry cache-slot addressing
+// ---------------------------------------------------------------------------
+//
+// The cache *key* for a non-registry tarball: an http(s) tarball is keyed by
+// its URL, a local `file:` tarball by its absolute path. Both share the
+// `cache_slot` sha256 primitive and live here — next to `commit_cache_dir_atomic`
+// (the slot *commit* protocol) — so all cache-slot concerns sit in one module
+// instead of the `file:` key living inside the http resolver. Consumed by
+// `super::http` and `super::file`, and re-exported to pm via the lib façade.
+
+/// `<prefix><16 hex>` from `sha256(bytes)`. Used by the two non-registry
+/// cache slots (`_http_<url>`, `_file_<path>`) to stay visually distinct
+/// from `<name>/<version>/` and 40-char git commit shas.
+#[cfg(feature = "http-tarball")]
+fn cache_slot(prefix: &str, bytes: &[u8]) -> String {
+    use std::fmt::Write as _;
+
+    let digest = Sha256::digest(bytes);
+    let mut out = String::with_capacity(prefix.len() + 16);
+    out.push_str(prefix);
+    for b in &digest[..8] {
+        let _ = write!(out, "{b:02x}");
+    }
+    out
+}
+
+/// Cache slot for an http(s) tarball, keyed on its URL.
+#[cfg(feature = "http-tarball")]
+pub fn http_cache_slot(url: &str) -> String {
+    cache_slot("_http_", url.as_bytes())
+}
+
+/// Cache slot for a local tarball keyed on its absolute path. The path is
+/// lexically normalized first so BFS and install hash to the same slot.
+///
+/// BFS resolves the tarball from a canonical absolute path
+/// (`base.join("/abs/foo.tgz")`), but install re-absolutizes a *root-relative*
+/// lockfile entry (`cwd.join("../../foo.tgz")`) whenever the tarball lives
+/// outside the project root. `Path::components()` collapses `.` and redundant
+/// separators but **keeps `..`**, so those two spellings of the same file would
+/// otherwise hash to different slots and the install-time lookup would miss.
+/// [`lexically_normalize`] collapses `..` so both spellings agree.
+#[cfg(feature = "http-tarball")]
+pub fn file_cache_slot(abs_path: &Path) -> String {
+    let normalized = lexically_normalize(abs_path);
+    cache_slot("_file_", normalized.as_os_str().as_encoded_bytes())
+}
+
+/// Collapse `.` and `..` purely lexically (no filesystem access, so it works
+/// for not-yet-existing paths and stays identical across BFS and install).
+///
+/// `..` pops the preceding normal component but never climbs past a root or
+/// prefix; a leading `..` in a relative path is preserved.
+///
+/// # Why lexical, not `fs::canonicalize`
+///
+/// This matches the semantics of Node's `path.resolve`/`path.normalize` — which
+/// is what npm uses to resolve `file:` specs — *not* `fs::realpath`. We only
+/// need BFS and install to agree on one canonical string for the cache key, and
+/// lexical collapsing achieves that without the costs of `fs::canonicalize`:
+/// the latter would require the tarball to already exist (breaking optional /
+/// not-yet-fetched deps), cost a syscall per package, and resolve symlinks —
+/// diverging from npm and changing the key based on FS state.
+///
+/// # Why hand-rolled
+///
+/// Rust std has no lexical `..`-collapse: `Path::components()` keeps `..` (the
+/// original bug), and `std::path::absolute` also deliberately keeps `..` on
+/// Unix. The `path-clean` crate exists in the tree but only as a transitive dep
+/// of the bundler half (swc/turbopack), not reachable from pm/ruborist, and it
+/// is not Windows-prefix aware. The `Component::Prefix` arm below handles
+/// Windows drive prefixes (e.g. `C:\`) so the win32 build stays correct even
+/// though no `file:`-tarball e2e currently exercises that target.
+#[cfg(feature = "http-tarball")]
+fn lexically_normalize(path: &Path) -> PathBuf {
+    use std::path::Component;
+
+    let mut stack: Vec<Component> = Vec::new();
+    for comp in path.components() {
+        match comp {
+            Component::CurDir => {}
+            Component::ParentDir => match stack.last() {
+                Some(Component::Normal(_)) => {
+                    stack.pop();
+                }
+                Some(Component::RootDir | Component::Prefix(_)) => {}
+                _ => stack.push(comp),
+            },
+            _ => stack.push(comp),
+        }
+    }
+    stack.iter().map(|c| c.as_os_str()).collect()
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
+
+    #[cfg(feature = "http-tarball")]
+    mod cache_slot {
+        use std::path::{Path, PathBuf};
+
+        use super::super::{file_cache_slot, http_cache_slot, lexically_normalize};
+
+        #[test]
+        fn http_slot_is_stable_and_url_specific() {
+            let a = http_cache_slot("https://example.com/foo.tgz");
+            let b = http_cache_slot("https://example.com/foo.tgz");
+            let c = http_cache_slot("https://example.com/foo.tgz?v=2");
+            assert_eq!(a, b);
+            assert_ne!(a, c);
+            assert!(a.starts_with("_http_"));
+            assert_eq!(a.len(), "_http_".len() + 16);
+        }
+
+        #[test]
+        fn file_slot_is_stable_across_dotdot_spellings() {
+            // BFS resolves from a canonical absolute path; install re-absolutizes
+            // a root-relative lockfile entry, which carries `..` when the tarball
+            // lives outside the project root. Both must hash to the same slot.
+            let canonical = Path::new("/tmp/pkgs/foo.tgz");
+            let via_dotdot = Path::new("/tmp/pkgs/app/../../pkgs/foo.tgz");
+            let via_curdir = Path::new("/tmp/pkgs/./foo.tgz");
+
+            assert_eq!(file_cache_slot(canonical), file_cache_slot(via_dotdot));
+            assert_eq!(file_cache_slot(canonical), file_cache_slot(via_curdir));
+            assert!(file_cache_slot(canonical).starts_with("_file_"));
+
+            // Distinct tarballs still get distinct slots.
+            assert_ne!(
+                file_cache_slot(canonical),
+                file_cache_slot(Path::new("/tmp/pkgs/bar.tgz"))
+            );
+        }
+
+        #[test]
+        fn lexically_normalize_collapses_dotdot_without_climbing_root() {
+            assert_eq!(
+                lexically_normalize(Path::new("/a/b/../c")),
+                PathBuf::from("/a/c")
+            );
+            // Excess `..` saturates at the root rather than escaping it.
+            assert_eq!(
+                lexically_normalize(Path::new("/a/../../../b")),
+                PathBuf::from("/b")
+            );
+            // Leading `..` in a relative path is preserved (nothing to pop).
+            assert_eq!(
+                lexically_normalize(Path::new("../a/b")),
+                PathBuf::from("../a/b")
+            );
+        }
+    }
 
     #[test]
     fn rejects_unsafe_package_names() {
